Refactor FormProcessor stages 14-17 into pipeline (#563 Sprint 1, PR 1b)
Complete the handle_submission_ajax split started in PR 1a. The persistence
(quiz + normal flows + the calculate_quiz_score / find_quiz_submission /
schedule-exception helpers), PDF generation, and success-response stages
move to SubmissionPersister, PdfStage, and SuccessResponder under
FreeFormCertificate\Frontend\Submission\. The persister/PDF/responder take
the SubmissionHandler as a constructor dependency.

The orchestrator is now a flat list of pipeline stages: guards + persister +
PDF + responder run in a single try; a SubmissionRejected is translated into
wp_send_json_error(); on success SuccessResponder populates $ctx->response,
emitted via wp_send_json_success(). FormProcessor drops from ~1047 to ~85
lines. Behavior-preserving: existing characterization tests + the per-guard
suite stay green (helpers' reflection-based tests repointed to
SubmissionPersister).

Also realigns develop's CHANGELOG with main: the released #564/#571 entries
move out of [Unreleased] into a [6.11.2] section (the manual release sync had
left them under [Unreleased] on develop).
